Preparing to connect to back-end
Made it so the user can't bet an amount that they don't have enough
coins for. If the user starts the game with less than 10 coins, the game
gives coins until he has 10 (enough for the lowest bet). Added the
ability for users to attach comments to questions they answer. Added a
new feature at the end of the game to view all answered questions as
well as user comments.

The connection code is ready, so once the back-end is complete we can
immediately hook it up and check for bugs.